[PIR][Dynamic Shape] Remove the insertino of tie_shape and shape.func ops (#59648)
Remove the insertino of tie_shape and shape.func ops
