+++ paddle/fluid/pybind/pir.cc
@@ -42,9 +42,9 @@
 #include "paddle/fluid/pir/transforms/fusion/fused_dropout_add_pass.h"
 #include "paddle/fluid/pir/transforms/fusion/fused_linear_param_grad_add_pass.h"
 #include "paddle/fluid/pir/transforms/fusion/fused_weight_only_linear_pass.h"
-#include "paddle/fluid/pir/transforms/infer_symbolic_shape_pass.h"
 #include "paddle/fluid/pir/transforms/inplace_pass.h"
 #include "paddle/fluid/pir/transforms/replace_fetch_with_shadow_output_pass.h"
+#include "paddle/fluid/pir/transforms/shape_optimization_pass.h"
 #include "paddle/fluid/pybind/control_flow_api.h"
 #include "paddle/phi/core/enforce.h"
 #include "paddle/pir/core/attribute.h"
@@ -1545,9 +1545,8 @@ void ApplyPirPass(Program &forward_program) {  // NOLINT
   bool has_dynamic_shape = HasDynamicShape(forward_program);
 
   auto shape_analysis =
-      has_dynamic_shape
-          ? std::make_shared<pir::MockShapeConstraintIRAnalysis>(ctx)
-          : nullptr;
+      has_dynamic_shape ? std::make_shared<pir::ShapeConstraintIRAnalysis>(ctx)
+                        : nullptr;
 
   pir::PassManager pass_manager(ctx);
   cinn::dialect::ir::PdOp2CinnOpConverter(&forward_program);
