+++ paddle/pir/dialect/shape/utils/shape_utils.h
@@ -71,7 +71,7 @@ class IR_API ShapeConstraintIRAnalysis : public ShapeAnalysis {
                       Value rhs,
                       std::vector<int> rhs_dim_idxs) override;
 
- protected:
+ private:
   // The operation this analysis runs on.
   ModuleOp m_;
   // The `SymbolicDimMgr` this analysis holds.
@@ -80,24 +80,18 @@ class IR_API ShapeConstraintIRAnalysis : public ShapeAnalysis {
   // dimension size of the memref value.
   std::unordered_map<Value, std::vector<shape::SymbolicDimOp>>
       value_to_sym_dims_;
-};
 
-class MockShapeConstraintIRAnalysis : public ShapeConstraintIRAnalysis {
  public:
-  explicit MockShapeConstraintIRAnalysis(
-      std::unique_ptr<pir::Program>&& program)
-      : ShapeConstraintIRAnalysis(program->module_op()),
-        program_(std::move(program)) {}
+  explicit ShapeConstraintIRAnalysis(std::shared_ptr<pir::Program>&& program)
+      : ShapeConstraintIRAnalysis(program->module_op()) {
+    program_ = std::move(program);
+  }
 
-  explicit MockShapeConstraintIRAnalysis(pir::IrContext* ctx)
-      : MockShapeConstraintIRAnalysis(std::make_unique<pir::Program>(ctx)) {}
-
-  MockShapeConstraintIRAnalysis(MockShapeConstraintIRAnalysis&& other) = delete;
-  MockShapeConstraintIRAnalysis(const MockShapeConstraintIRAnalysis& other) =
-      delete;
+  explicit ShapeConstraintIRAnalysis(pir::IrContext* ctx)
+      : ShapeConstraintIRAnalysis(std::make_shared<pir::Program>(ctx)) {}
 
  private:
-  std::unique_ptr<pir::Program> program_;
+  std::shared_ptr<pir::Program> program_;
 };
 
 class IR_API ShapeAnalysisManager {
