+++ test/cpp/pir/shape_dialect/CMakeLists.txt
@@ -16,17 +16,20 @@ paddle_test(
   pir
   gtest)
 
-paddle_test(
-  shape_optimization_test
-  SRCS
-  shape_optimization_test.cc
-  DEPS
-  gtest
-  op_dialect_vjp
-  pir)
-
-set_tests_properties(shape_optimization_test
-                     PROPERTIES ENVIRONMENT "FLAGS_enable_pir_in_executor=true")
+if(WITH_CINN)
+  paddle_test(
+    shape_optimization_test
+    SRCS
+    shape_optimization_test.cc
+    DEPS
+    gtest
+    op_dialect_vjp
+    pir
+    pir_transforms)
+  set_tests_properties(
+    shape_optimization_test PROPERTIES ENVIRONMENT
+                                       "FLAGS_enable_pir_in_executor=true")
+endif()
 
 if(WITH_ONNXRUNTIME AND WIN32)
   # Copy onnxruntime for some c++ test in Windows, since the test will
